+++ slack-plugin/src/main/kotlin/slack/gradle/StandardProjectConfigurations.kt
@@ -660,6 +660,18 @@ internal class StandardProjectConfigurations(
                 ?.contains(builder.name)
                 ?: true
           builder.enableAndroidTest = variantEnabled
+          if (variantEnabled) {
+            // Ensure there's a manifest file present and has its debuggable flag set correctly
+            if (slackProperties.strictMode && slackProperties.strictValidateAndroidTestManifest) {
+              val manifest = project.file("src/androidTest/AndroidManifest.xml")
+              check(manifest.exists()) {
+                "AndroidManifest.xml is missing from src/androidTest. Ensure it exists and also is set to debuggable!"
+              }
+              check(manifest.readText().contains("android:debuggable=\"true\"")) {
+                "AndroidManifest.xml in src/androidTest is missing the debuggable flag! Ensure it is set to 'android:debuggable=\"true\"'"
+              }
+            }
+          }
         }
 
         // Contribute these libraries to Fladle if they opt into it
@@ -859,7 +871,8 @@ internal class StandardProjectConfigurations(
               )
             }CompilationMarker
             ```
-            """.trimIndent()
+            """
+              .trimIndent()
           )
         }
       }